test(e2e): admin-insights spec with redirect + 4-section + nav coverage
